+++ CHANGES.rst
@@ -4,6 +4,9 @@
 Bug Fixes
 ---------
 
+- Histogram in Plot Options now uses random sampling to better
+  represent the data without sacrificing performance. [#2771]
+
 Cubeviz
 ^^^^^^^